[update-image-editor-code] [Image] | (DX) | Modernize the Image widget edtior code
- Remove unused code
- Break out the extra settings into another component (functional component :) )

NOTE: Changes to style are part of a different ticket. This is just moving code around.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3383

Test plan:
Storybook
- Go to `/?path=/story/editors-editorpage--demo`
- Create an image widget
- Paste an image URL (I'm using `https://cdn.kastatic.org/ka-content-images/61831c1329dbc32036d7dd0d03e06e7e2c622718.jpg`)
- Confirm that changing the caption changes the caption in the preview
- Change the alt and use the browser inspector to confirm it updated in the preview
