+++ packages/perseus-editor/src/all-editors.ts
@@ -11,7 +11,7 @@ import GradedGroupSetEditor from "./widgets/graded-group-set-editor";
 import GrapherEditor from "./widgets/grapher-editor";
 import GroupEditor from "./widgets/group-editor";
 import IframeEditor from "./widgets/iframe-editor";
-import ImageEditor from "./widgets/image-editor";
+import ImageEditor from "./widgets/image-editor/image-editor";
 import InputNumberEditor from "./widgets/input-number-editor";
 import InteractionEditor from "./widgets/interaction-editor";
 import InteractiveGraphEditor from "./widgets/interactive-graph-editor/interactive-graph-editor";
